Added context memory
Chat history is store in ContentManager per user. Chat history is sent to the model with each request, turning the interaction with model into a chat (as opposed to series of context free question-answer series). To clear the context, user mst send message starting with `/clear` to the Telegram bot.